Добавил скрипты SQL, настройки окружения.

+++ db/scripts/004_ddl_create_price_history_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE price_history(
-   id SERIAL PRIMARY KEY,
+   id BIGSERIAL PRIMARY KEY,
    before BIGINT not null,
    after BIGINT not null,
    created TIMESTAMP WITHOUT TIME ZONE DEFAULT now(),
-   auto_post_id INT REFERENCES auto_post(id)
+   auto_post_id BIGINT REFERENCES auto_post(id)
 );
